MueLu: initial clean-up of ImageBasedPFactory
There are things that needs further attention, I am not sure that the logic around the "greyscale" input data is correct.
It would also be nice to review the use of the coordinates and other input/ouput parameters.
The new test is derived from the structured test but also generates image based specific data.
It is currently guarded using the EXPERIMENTAL configuration flag.